feature: validated codex plugin
