+++ web-ui/packages/codex-mcp/__tests__/smoke.test.ts
@@ -3,7 +3,7 @@ import path from 'path';
 
 describe('Codex plugin manifest', () => {
   it('should exist and be valid JSON', () => {
-    const manifestPath = path.join(__dirname, '../.codex-plugin/plugin.json');
+    const manifestPath = path.join(__dirname, '../src/.codex-plugin/plugin.json');
     expect(fs.existsSync(manifestPath)).toBe(true);
     const manifest = JSON.parse(fs.readFileSync(manifestPath, 'utf-8'));
     expect(manifest).toHaveProperty('name');
