--- web-ui/packages/codex-mcp/src/scripts/oauth-mcp-wrapper.ts
@@ -1,4 +1,0 @@
-// TODO: Port logic from oauth-mcp-wrapper.mjs to TypeScript
-// This is a placeholder. The full port should replicate all CLI and helper logic.
-
-console.log('oauth-mcp-wrapper.ts: Not yet implemented. Port logic from oauth-mcp-wrapper.mjs.');
