--- web-ui/packages/codex-mcp/src/scripts/runtime-utils.test.ts
@@ -1,81 +1,0 @@
-import assert from "node:assert/strict";
-import test from "node:test";
-import {
-  appSessionCookieHeader,
-  backoffDelayMs,
-  // fetchWithPolicy, // Not ported
-  isAuthenticatedSessionResult,
-  isUsableCachedAppSession,
-  isUsableCachedToken,
-  parseNumber,
-  // resolveEndpoint, // Not ported
-  // warnIfInsecureUrl // Not ported
-} from "./runtime-utils";
-
-test("parseNumber falls back and clamps to the minimum", () => {
-  assert.equal(parseNumber(undefined, 15, 1), 15);
-  assert.equal(parseNumber("-5", 15, 0), 0);
-  assert.equal(parseNumber("25", 15, 0), 25);
-});
-
-test("backoffDelayMs grows exponentially", () => {
-  assert.equal(backoffDelayMs(0, 250), 250);
-  assert.equal(backoffDelayMs(2, 250), 1000);
-});
-
-test("isUsableCachedToken honors the expiry skew", () => {
-  assert.equal(
-    isUsableCachedToken({ access_token: "token", expires_at: Date.now() + 120000 }, 60000),
-    true
-  );
-  assert.equal(
-    isUsableCachedToken({ access_token: "token", expires_at: Date.now() + 30000 }, 60000),
-    false
-  );
-});
-
-test("isUsableCachedAppSession requires a current wrapped session cookie", () => {
-  assert.equal(
-    isUsableCachedAppSession({
-      app_session: {
-        token: "wrapped",
-        cookie_name: "authjs.session-token",
-        expires_at: Date.now() + 120000
-      }
-    }, 60000),
-    true
-  );
-  assert.equal(
-    isUsableCachedAppSession({
-      app_session: {
-        token: "wrapped",
-        cookie_name: "authjs.session-token",
-        expires_at: Date.now() + 30000
-      }
-    }, 60000),
-    false
-  );
-});
-
-test("appSessionCookieHeader formats the wrapped Auth.js cookie", () => {
-  assert.equal(
-    appSessionCookieHeader({ token: "wrapped", cookie_name: "authjs.session-token" }),
-    "authjs.session-token=wrapped"
-  );
-  assert.equal(appSessionCookieHeader({ token: "wrapped" }), undefined);
-});
-
-test("isAuthenticatedSessionResult requires an authenticated app session body", () => {
-  assert.equal(
-    isAuthenticatedSessionResult({ response: { ok: true }, body: { status: "authenticated" } }),
-    true
-  );
-  assert.equal(
-    isAuthenticatedSessionResult({ response: { ok: true }, body: { status: "unauthenticated" } }),
-    false
-  );
-  assert.equal(
-    isAuthenticatedSessionResult({ response: { ok: false }, body: { status: "authenticated" } }),
-    false
-  );
-});
