+++ web-ui/packages/codex-mcp/src/scripts/smoke-list-abilities-sse.mjs
