--- web-ui/packages/codex-mcp/src/scripts/smoke-list-abilities-sse.ts
@@ -1,2 +1,0 @@
-// TODO: Port logic from smoke-list-abilities-sse.mjs to TypeScript
-console.log('smoke-list-abilities-sse.ts: Not yet implemented. Port logic from smoke-list-abilities-sse.mjs.');
