+++ web-ui/packages/codex-mcp/src/scripts/smoke-list-resources-sse.mjs
