--- web-ui/packages/codex-mcp/src/scripts/smoke-list-resources-sse.ts
@@ -1,2 +1,0 @@
-// TODO: Port logic from smoke-list-resources-sse.mjs to TypeScript
-console.log('smoke-list-resources-sse.ts: Not yet implemented. Port logic from smoke-list-resources-sse.mjs.');
